+++ R/nfsa_sequence_T0800_reg.R
@@ -33,7 +33,6 @@
 #' @export
 nfsa_sequence_T0800_reg <- function(country,
                                 time_min = 2020,
-                                type = "values",
                                 template = here("assets","seq_accounts_T0800_reg.xlsx"),
                                 output_sel = here("output","sequence")) {
 library(tidyverse)