+++ R/nfsa_sequence_T0801_reg.R
@@ -34,7 +34,6 @@
 #' @export
 nfsa_sequence_T0801_reg <- function(country,
                                 time_min = "2020-Q1",
-                                type = "values",
                                 template = here("assets","seq_accounts_T0801_reg.xlsx"),
                                 output_sel = here("output","sequence")) {